PriaccEnterpriseAI/ai-assisstant/requirements.txt
fastapi
uvicorn
pydantic
langchain
openai
python-dotenv
langcahin_openai